refactor(evm): consolidate client-side createPermit2Payload into shared helper
