+++ typescript/packages/mechanisms/evm/src/shared/permit2-helpers.ts
@@ -1,5 +1,6 @@
 import {
   PaymentPayload,
+  PaymentPayloadResult,
   PaymentRequirements,
   FacilitatorContext,
   SettleResponse,
@@ -15,7 +16,8 @@ import {
   type Erc20ApprovalGasSponsoringSigner,
 } from "../exact/extensions";
 import { getAddress, encodeFunctionData } from "viem";
-import { PERMIT2_ADDRESS, erc20AllowanceAbi } from "../constants";
+import { PERMIT2_ADDRESS, erc20AllowanceAbi, permit2WitnessTypes } from "../constants";
+import { createPermit2Nonce, getEvmChainId } from "../utils";
 import {
   ErrPermit2612AmountMismatch,
   ErrPermit2InvalidAmount,
@@ -25,8 +27,8 @@ import {
   ErrPermit2InvalidSignature,
   ErrPermit2PaymentTooEarly,
 } from "../exact/facilitator/errors";
-import { FacilitatorEvmSigner } from "../signer";
-import { ExactPermit2Payload, UptoPermit2Payload } from "../types";
+import { ClientEvmSigner, FacilitatorEvmSigner } from "../signer";
+import { ExactPermit2Payload, Permit2Authorization, UptoPermit2Payload } from "../types";
 import { validateErc20ApprovalForPayment } from "../exact/facilitator/erc20approval";
 
 /**
@@ -443,4 +445,95 @@ export async function settlePermit2Direct(
   } catch (error) {
     return mapSettleError(error, payload, payer);
   }
+}
+
+// ---------------------------------------------------------------------------
+// Client-side helpers
+// ---------------------------------------------------------------------------
+
+/**
+ * Creates a Permit2 payload for any scheme (exact or upto).
+ * The only scheme-specific input is the proxy address used as the spender.
+ *
+ * @param proxyAddress - The x402 proxy contract address to set as spender
+ * @param signer - The EVM client signer
+ * @param x402Version - The x402 protocol version
+ * @param paymentRequirements - The payment requirements
+ * @returns Promise resolving to a payment payload result
+ */
+export async function createPermit2PayloadForProxy(
+  proxyAddress: `0x${string}`,
+  signer: ClientEvmSigner,
+  x402Version: number,
+  paymentRequirements: PaymentRequirements,
+): Promise<PaymentPayloadResult> {
+  const now = Math.floor(Date.now() / 1000);
+  const nonce = createPermit2Nonce();
+
+  // Lower time bound - allow some clock skew
+  const validAfter = (now - 600).toString();
+  // Upper time bound is enforced by Permit2's deadline field
+  const deadline = (now + paymentRequirements.maxTimeoutSeconds).toString();
+
+  const permit2Authorization: Permit2Authorization & { from: `0x${string}` } = {
+    from: signer.address,
+    permitted: {
+      token: getAddress(paymentRequirements.asset),
+      amount: paymentRequirements.amount,
+    },
+    spender: proxyAddress,
+    nonce,
+    deadline,
+    witness: {
+      to: getAddress(paymentRequirements.payTo),
+      validAfter,
+    },
+  };
+
+  const signature = await signPermit2Authorization(
+    signer,
+    permit2Authorization,
+    paymentRequirements,
+  );
+
+  return {
+    x402Version,
+    payload: { signature, permit2Authorization },
+  };
+}
+
+/**
+ * Signs a Permit2 authorization using EIP-712 with witness data.
+ * The signature authorizes the proxy contract to transfer tokens on behalf of the signer.
+ *
+ * @param signer - The EVM client signer
+ * @param permit2Authorization - The Permit2 authorization parameters
+ * @param requirements - The payment requirements
+ * @returns Promise resolving to the hex-encoded signature
+ */
+async function signPermit2Authorization(
+  signer: ClientEvmSigner,
+  permit2Authorization: Permit2Authorization & { from: `0x${string}` },
+  requirements: PaymentRequirements,
+): Promise<`0x${string}`> {
+  const chainId = getEvmChainId(requirements.network);
+
+  return await signer.signTypedData({
+    domain: { name: "Permit2", chainId, verifyingContract: PERMIT2_ADDRESS },
+    types: permit2WitnessTypes,
+    primaryType: "PermitWitnessTransferFrom",
+    message: {
+      permitted: {
+        token: getAddress(permit2Authorization.permitted.token),
+        amount: BigInt(permit2Authorization.permitted.amount),
+      },
+      spender: getAddress(permit2Authorization.spender),
+      nonce: BigInt(permit2Authorization.nonce),
+      deadline: BigInt(permit2Authorization.deadline),
+      witness: {
+        to: getAddress(permit2Authorization.witness.to),
+        validAfter: BigInt(permit2Authorization.witness.validAfter),
+      },
+    },
+  });
 }
